Support decimal coupon percent_off end to end
percent_off now flows as float64 through coupon evidence, discount
metadata snapshots, and renewal invoices, matching stripe-node v22's
number contract (e.g. 12.5). Percent discounts round half away from
zero; existing integer-percent cases are unchanged. The SDK smoke
covers decimal coupon round-trip and checkout discount amounts.

+++ internal/api/api_test.go
@@ -3495,9 +3495,9 @@ func TestExpandedStripeSurfaceSimulations(t *testing.T) {
 	})
 
 	coupon := postForm[struct {
-		ID         string `json:"id"`
-		Object     string `json:"object"`
-		PercentOff int64  `json:"percent_off"`
+		ID         string  `json:"id"`
+		Object     string  `json:"object"`
+		PercentOff float64 `json:"percent_off"`
 	}](t, handler, "/v1/coupons", url.Values{
 		"id":          {"coupon_launch"},
 		"percent_off": {"25"},
